Add mice
Well, they look like a boring icon. But they have sort of the right
properties.

I introduced a CollisionController to do the heavy lifting of keeping
track of when mice get trapped in a bubble. Seems to work ok, at least
on a desktop computer. Though I worry that it'll get slow.